Controladores y Recursos API funcionando

+++ database/migrations/2025_12_03_000006_create_criterios_evaluacion_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('criterios_evaluacion', function (Blueprint $table) {
 
             $table->id()->primary();
-            $table->integer('resultado_aprendizaje_id')->nullable();
+            $table->unsignedBigInteger('resultado_aprendizaje_id')->nullable();
                 $table->foreign('resultado_aprendizaje_id')->references('id')->on('resultados_aprendizaje')->onDelete('cascade');
             $table->string('codigo',50);
             $table->string('descripcion');
